+++ rust/stackablectl/CHANGELOG.md
@@ -4,11 +4,16 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Add shell completions for Nushell and Elvish ([#337]).
+
 ### Fixed
 
 - Sort operator versions by semver version instead of alphabetically ([#336]).
 
 [#336]: https://github.com/stackabletech/stackable-cockpit/pull/336
+[#337]: https://github.com/stackabletech/stackable-cockpit/pull/337
 
 ## [24.11.0] - 2024-11-18